requirements.txt
Flask==3.0.3
flask-wtf==1.2.1
Werkzeug==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
SQLAlchemy==2.0.32
alembic==1.13.2
cryptography==43.0.0
python-dotenv==1.0.1
requests==2.32.3
APScheduler==3.10.4
hvac==2.3.0
boto3==1.34.162
authlib==1.3.1
pytest==8.3.2
